mysite/pyproject.toml
[project]
name = "mysite"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "django-taggit>=6.1.0",
    "markdown>=3.10",
    "psycopg2-binary>=2.9.11",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 100
